chore(orchestrator): make OrchestratorFormContext singleton in dynamic plugins
