+++ workspaces/orchestrator/plugins/orchestrator-form-api/src/DefaultFormApi.tsx
@@ -14,15 +14,28 @@
  * limitations under the License.
  */
 
-import React from 'react';
+import React, { Context } from 'react';
 
 import {
   FormDecoratorProps,
   OrchestratorFormApi,
+  OrchestratorFormContextProps,
   OrchestratorFormDecorator,
 } from '@red-hat-developer-hub/backstage-plugin-orchestrator-form-api';
 
 class DefaultFormApi implements OrchestratorFormApi {
+  private readonly context: Context<OrchestratorFormContextProps | null>;
+
+  public constructor() {
+    this.context = React.createContext<OrchestratorFormContextProps | null>(
+      null,
+    );
+  }
+
+  getFormContext(): React.Context<OrchestratorFormContextProps | null> {
+    return this.context;
+  }
+
   getFormDecorator(): OrchestratorFormDecorator {
     // eslint-disable-next-line no-console
     console.log(
